requirements.txt
jsonschema
pyyaml>=5.1
future
